Move convert_mlflow_uri_to_litellm to model_utils, remove duplicate from dspy_utils
The convert_mlflow_uri_to_litellm function was defined locally in
dspy_utils.py. Move it to the canonical model_utils.py location
alongside _parse_model_uri which it depends on, and update all
imports across dspy_utils, memalign, and their tests.
