+++ tests/genai/judges/optimizers/test_dspy_utils.py
@@ -11,7 +11,6 @@ from mlflow.genai.judges.optimizers.dspy_utils import (
     append_input_fields_section,
     construct_dspy_lm,
     convert_litellm_to_mlflow_uri,
-    convert_mlflow_uri_to_litellm,
     create_dspy_signature,
     format_demos_as_examples,
     trace_to_dspy_example,
@@ -21,6 +20,7 @@ from mlflow.genai.utils.trace_utils import (
     extract_request_from_trace,
     extract_response_from_trace,
 )
+from mlflow.metrics.genai.model_utils import convert_mlflow_uri_to_litellm
 
 from tests.genai.judges.optimizers.conftest import MockJudge
